biometric-based electronic voting
this program is under construction
database -> MySql
code editor -> Visual Studio Code
used

The purpose of this program is biometric-based on the web.
electronic voting

While registering, it asks for your e-mail address and sends you a code.
If you enter the code sent correctly, it requests and saves your face data.
Afterwards, when you enter the correct ID number and password while logging in, you will be asked.
asks for your face data and sees whether it matches the face you provided when registering.
checking is true, login is successful